MYSQL-267: Add user@host to binlog event metadata
This feature adds the functionality to emit user@host information when
writing an event to binlog. This will be read by mysqlbinlog. We use
the existing Q_INVOKER field in query event header which is used by
the replicating slave to read user info of REVOKE and GRANT commands.

– A new option binlog_write_user_info (default is false)
– When above option is set on master binlog event will have an added
user@host field:
– If the option is enabled on slave and the slave is setup for
binlogging then user info won't be displayed when slave SQL thread
executes a query except if the command was REVOKE or GRANT. And in
the case of REVOKE or GRANT the user info displayed won't be of the
slave thread but the actual invoker of the command on the master.
– Direct queries on slave will have the info displayed.