BUG#17066269: AUTO_INC VALUE NOT PROPERLY GENERATED WITH RBR AND AUTO_INC COLUMN ONLY ON SLAVE
In RBR, if the slave's table as one additional auto_inc column,
then, it will insert the value 0 instead of generating the next
auto_inc number.

We fix this by checking that if an auto_inc extra column exists,
when compared to column data of the row event, we explicitly set
it to NULL and flag the engine that a nulled auto_inc column will
be inserted.
--BZR--
revision-id: luis.soares@oracle.com-20131218111724-t528fgvx3blgopas
property-audit-revid: luis.soares@oracle.com-20131218111724-k515ruvbhb60p9m6
property-branch-nick: mysql-5.5
testament3-sha1: 25b32dd5137163e2e32074519c1cae84f107cbd4