Bug#11759519 - INFINITE HANG WITH 100% CPU USAGE WITH LOAD DATA                LOCAL AND IMPORT ERRORS
Description:
-----------
This bug happens due to the fact that current algorithm is designed
that in the case of LOCAL load of data, in case of the error, the
remaining part of the file is read in order to return the proper
error message to the client side.

But, the problem with current implementation is that data stream
for the client side is cleared only in the case where line delimiters
exist, which is not a case with, for example fixed width
fields.

Fix:
----
Ported patch provided by Sinisa Milivojevic n bug report for this
issue to 5.5+ versions.

As part of this patch code is changed to clear the data stream
by calling new member function "READ_INFO::skip_data_till_eof".

--BZR--
revision-id: praveenkumar.hulakund@oracle.com-20140319070030-702yfk1n7tuy7223
property-audit-revid: praveenkumar.hulakund@oracle.com-20140319070030-0bc5kvg7sgx0355d
property-branch-nick: mysql-5.5
testament3-sha1: e1b894e3c9dc1ef2cd5021bb6e4c00f8c61de8b2