BUG#18080920: CRASH; MY_REALLOC_STR DEREFERENCES NEGATIVE VALUE               INTO CLIENT_ERRORS ARRAY
Analysis:
--------
The client may crash while executing a statement due to
the missing mapping of the server error to it's equivalent
client error.

When trying to reallocate memory for the packet buffer, if
the system is out of memory or the packet buffer is large,
the server errors 'ER_OUT_OF_RESOURCES' or 'ER_PACKET_TOO_LARGE'
is returned respectively. The client error number calculated is
negative and when trying to dereference the array of client 
error messages with the calculated error number, the client
crashes.

Fix:
----
Map the server error returned to it's equivalent client error
prior to dereferencing the array of client error messages.

Note: Test case is not added since it is difficult to simulate
the error condition.
--BZR--
revision-id: nisha.gopalakrishnan@oracle.com-20140424040021-u7aqw422x0rt5eut
property-audit-revid: nisha.gopalakrishnan@oracle.com-20140424040021-vyfi5o50ghhpgi2o
property-branch-nick: mysql-5.5-18080920
testament3-sha1: 3edda95258521e8ce2ec8c1b64ecd8a2f8d4fb44