MYSQL-307: Increase maximum user name length to 32 characters
Incorporated Facebook's patch to increase username length to 32
characters, as well as fixing a few regression tests that were affected.

Note: the data dictionary on hosts running this change will need to
be updated in mysql.user, otherwise in the column User, they will try
to cram a 32-character string into a space allocated for 16 characters
and it will be truncated.