Bug #18163964 PASSWORD IS VISIBLE WHILE CHANGING IT FROM               MYSQLADMIN IN PROCESSES LIST
Description: Checking the process status (with ps -ef)  
while executing "mysqladmin" with old password and new 
password via command-line will show the new password in the
process list sporadically.

Analysis: The old password is being masked by "mysqladmin".
So masking the new password in the similar manner would 
reduce hitting the bug. But this would not completely fix
the bug, because if "ps -ef " command hits the mysqladmin
before it masks the passwords it will show both the old and
new passwords in the process list. But the chances of
hitting this is very less.

Fix: The new password also masked in the similar manner
that of the --password argument.
--BZR--
revision-id: arun.kuruvila@oracle.com-20140516034639-bwqyu4g8buz2tgsj
property-audit-revid: arun.kuruvila@oracle.com-20140516034639-uhb7htota0hcodjv
property-branch-nick: mysql-5.5
testament3-sha1: 9d1f4ad4b363d780cf3d41b9a05f598d032e77e8