Bug#18469276: MOD FOR SMALL DECIMALS FAILS
Problem:
If leading zeroes of fractional part of a decimal
number exceeds 45, mod operation on the same fails.
      
Analysis:
Currently there is a miscalcultion of fractional
part for very small decimals in do_div_mod.
      
For ex:
For 0.000(45 times).....3
length of the integer part becomes -5 (for a length of one,
buffer can hold 9 digits. Since number of zeroes are 45, integer
part becomes 5) and it is negative because of the leading
zeroes present in the fractional part.
Fractional part is the number of digits present after the
point which is 46 and therefore rounded off to the nearest 9
multiple which is 54. So the length of the resulting fractional
part becomes 6.
      
Because of this, the combined length of integer part and fractional
part exceeds the max length allocated which is 9 and thereby failing.
      
Solution:
In case of negative integer value, it indicates there are
leading zeroes in fractional part. As a result stop1 pointer 
should be set not just based on frac0 but also intg0. This is
because the detination buffer will be filled with 0's for the length
of intg0.
--BZR--
revision-id: chaithra.gopalareddy@oracle.com-20140703084202-g68kysanck2k3ltn
property-audit-revid: chaithra.gopalareddy@oracle.com-20140703084202-0iy67hhi3b9wueu4
property-branch-nick: mysql-5.5
property-file-info: ld7:file_id62:sp1f-decimal.c-20041018120639-qqm4oeadwhbdhuxgwlc657il4tu44n6m7:message63:Calculate stop1 pointer based on the length of intg0 and frac0.4:path17:strings/decimal.cee
testament3-sha1: 914763d53ed98e3d5087133bc321d891bef2fb81